requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-huggingface
faiss-cpu
numpy
sentence-transformers
python-dotenv
google-generativeai
fastapi
uvicorn[standard]
